refactor(validator): domain 패키지 내 validator 패키지 생성
