+++ src/main/java/domain/GameTrialCount.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.validator.GameTrialCountValidator;
+
 public class GameTrialCount {
 
     int gameTrial;
