+++ src/main/java/domain/validator/GameTrialCountValidator.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.validator;
 
 import validation.ErrorMessage;
